Renamed components and files so that they are all descriptive and relevant. #6
