+++ client/components/campground-list.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-export default function CampgroundListItem(props) {
+export default function CampgroundList(props) {
 
   return (
     <div className='container-fluid'>
